Fixed markdown for generics
