+++ package.json
@@ -5,9 +5,9 @@
   "scripts": {
     "build:copy-files": "ts-node -T scripts/copy-files.ts",
     "build:types": "ts-node -T scripts/add-types-to-readme.ts",
-    "build:make-toc": "markdown-toc -i README.md",
+    "build:toc": "markdown-toc -i README.md",
     "build:tsc": "rimraf dist && tsc --build tsconfig.json",
-    "build": "npm run build:tsc && npm run build:types && npm run build:make-toc && npm run build:copy-files",
+    "build": "npm run build:tsc && npm run build:types && npm run build:toc && npm run build:copy-files",
     "test": "jest",
     "test:watch": "jest --watch"
   },
