+++ src/helpers/makeSessionConfigObject.ts
@@ -1,17 +1,5 @@
-import { MemoryStore, SessionOptions, Store } from "express-session"
-
-export interface IMakeSessionConfigObjectOptions {
-	/**
-	 * A unique string that is used to sign the session id.
-	 * This MUST NOT be shared with any other application.
-	 */
-	secret: string
-	/**
-	 * A store instance where session data will be stored.
-	 * You MUST provide this otherwise express-session will default to using the insecure memory store.
-	 */
-	store: Store | MemoryStore
-}
+import { SessionOptions } from "express-session"
+import { IMakeSessionConfigObjectOptions } from "../interfaces/IMakeSessionConfigObject"
 
 const recommendedSessionOptions = {
 	name: "veracity.session",
